requirements.txt
docopt==0.6.2
voluptuous==0.8.4
six==1.8.0
ExifRead==1.4.2
PyYAML==3.11
-e git+https://github.com/scikit-image/scikit-image.git@master#egg=scikit-image
